development/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "ML API using FastAPI"
authors = ["JordanMarcelino <jhordan.marcelino@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.110.0"
slowapi = "^0.1.9"
sqlmodel = "^0.0.16"
pydantic-settings = "^2.2.1"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
scikit-learn = "^1.4.1.post1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
loguru = "^0.7.2"
psycopg2 = "^2.9.9"
bcrypt = "^3.2.0"
pydantic = {extras = ["email"], version = "^2.6.3"}


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
pytest-cov = "^4.1.0"
httpx = "^0.27.0"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore",
    "ignore::UserWarning",
    # note the use of single quote below to denote "raw" strings in TOML
    'ignore:function ham\(\) is deprecated:DeprecationWarning',
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
